Add files via upload
This program calculates the risk map of the building stoke of an area for a given seismic hazard by applying a simplidied mechanical method. The risk is defined as the probability of a certain level of damage to be exceeded. Four levels of damage are examined, namely P1,P2,P3 and P4. The level of damage Pi increases with increasing i=1 to 4. The probability takes valuse from zero (totally unlikely) to one (cerntainty). 

Input parameters of the model: 
• Structural data (typology, construction period, use of building, post-earthquake damage data, detection of different construction phases, etc) 
• Geometrical data (total height, number of storeys, perimeter, footprint area, average bearing wall thickness for masonry structures, Plan view at least of the ground-floor, etc)
* The Seismic hazard data: The seismic hazard data include different seismic acceleration spectra at a grid of points within the area of interest. 

 An acceleration spectrum corresponds at each point which has resulted from a seismic hazard analysis taking into account the seismic sources and the local soil effect (e.g., Tselentis et al 2010). 


STABLE (STructural stABiLity risk assEssment, 2018-2022, Horizon2020, MSCA - RISE - grant agreement n. 823966),